(PC-41171)[PRO] refactor: move listOffersVenueV2Factory into common offer factories
