+++ pro/src/commons/utils/factories/individualApiFactories.ts
@@ -37,7 +37,7 @@ import { CATEGORY_STATUS } from '@/commons/core/Offers/constants'
 import type { StocksEvent } from '@/pages/IndividualOffer/IndividualOfferTimetable/components/StocksCalendar/form/types'
 
 import type { PartialExcept } from '../types'
-import { listOffersVenueV2Factory } from './collectiveApiFactories'
+import { listOffersVenueV2Factory } from './commonOffersApiFactories'
 
 let offerId = 1
 let stockId = 1
